Fix exception when CommentsTemplate doesn't exist
When using a template in the theme, the cache key when looking for the CommentsTemplate.html file is an absolute path with "backwards" slashes, e.g. C:/inetpub/wwwroot/example.com/Portals/_default/Skins/example/Templates/Blog/example/CommentsTemplate.html

When that doesn't exist, it was passed to HostingEnvironment.MapPath, which gave an exception at meeting such a path ("The relative virtual path 'C:/inetpub/…/CommentsTemplate.html' is not allowed here").

This commit avoid this by never passing a rooted path to HostingEnvironment.MapPath, and only doing the File.Exists check once.