Make list images link to posts
Many popular blogs (ex: TechCrunch, Mashable) and news sites allow the
user to click on the listi image (thumbnail) to link to the full post,
with the same results as if the user clicked on the title text.

This update to the _default skin adds the full post's link reference to
the list image by wrapping the HTML <figure> tag.

A related change in the ListImage.html file: the HTML <picture> tag is
removed, since it is not used/needed in this skin. The picture tag is
designed/intended to specify multiple images of varying sizes, allowing
the broswer to choose the most appropriate image to display based upon
the browser's available window size. (A description with example can be
found at https://www.w3schools.com/tags/tag_picture.asp .) Since the
DNN-Blog only allows a single image for the post, the picture tag is
unnecessary.